requirements.txt
flask
matplotlib
numpy
pandas
scikit-learn
gunicorn
Flask-MySQLdb
python-dotenv